Change name FileSinkSVG and FileSinkSVG2
