+++ src/org/graphstream/stream/file/FileSinkFactory.java
@@ -56,7 +56,7 @@ public class FileSinkFactory {
 		ext2sink.put("gml", FileSinkGML.class);
 		ext2sink.put("graphml", FileSinkGraphML.class);
 		ext2sink.put("dot", FileSinkDOT.class);
-		ext2sink.put("svg", FileSinkSVG2.class);
+		ext2sink.put("svg", FileSinkSVG.class);
 		ext2sink.put("pgf", FileSinkTikZ.class);
 		ext2sink.put("tikz", FileSinkTikZ.class);
 		ext2sink.put("tex", FileSinkTikZ.class);
